apps/api/pyproject.toml
[project]
name = "manimations"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "google-genai>=1.57.0",
    "manim>=0.19.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.40.0",
]
